+++ automatix/bundlewrap.py
@@ -3,7 +3,7 @@ from bundlewrap.group import Group
 from bundlewrap.node import Node
 from bundlewrap.repo import Repository
 
-from .command import Command
+from .command import Command, PA
 
 
 class AutomatixBwRepo(Repository):
@@ -23,7 +23,8 @@ class BWCommand(Command):
                         self.env.config['bw_repo'].get_group(value)
                     except NoSuchGroup:
                         self.env.LOG.warning(f'"{value}" is neither a BW node nor a BW group')
-        locale_vars['a_vars'] = self.env.vars
+        locale_vars['a_vars'] = self.env.vars  # deprecated, for backwards compatibility
+        locale_vars['VARS'] = self.env.vars
         locale_vars['NODES'] = BWNodesWrapper(repo=self.env.config['bw_repo'], systems=self.env.systems)
         return locale_vars
 
@@ -55,9 +56,14 @@ class BWCommand(Command):
             if self.env.cmd_args.force:
                 return
 
-            err_answer = self._ask_user(question='[PF] What should I do?', allowed_options=['p', 'T', 'v', 'r', 'a'])
-            # answers 'a' and 'c' are handled by _ask_user, 'p' means just pass
-            if err_answer == 'r':
+            err_answer = self._ask_user(
+                question='[PF] What should I do?',
+                allowed_options=[PA.proceed, PA.terminal, PA.variables, PA.retry, PA.abort],
+            )
+            # _ask_user handles are answers but PA.retry, PA.skip, PA.proceed
+            # PA.skip is not in the allowed options'
+            # PA.proceed means 'proceed' so we can just go on
+            if err_answer == PA.retry.answer:
                 return self._remote_bw_group_action(node=node)