+++ automatix/config_test.py
@@ -50,7 +50,11 @@ def test__check_deprecated_syntax__global(caplog):
 
 def test__check_deprecated_syntax__vars(caplog):
     check_deprecated_syntax(ckey='python', entry='vars["myvar"] = xyz', script={}, prefix='[pipeline:7]')
-    assert '[pipeline:7] Using "vars["myvar"]" does not work any longer. Use "a_vars["myvar"]" instead.' in caplog.text
+    assert '[pipeline:7] Using "vars["myvar"]" does not work any longer. Use "VARS.myvar" instead.' in caplog.text
+
+def test__check_deprecated_syntax__a_vars(caplog):
+    check_deprecated_syntax(ckey='python', entry='a_vars["myvar"] = xyz', script={}, prefix='[pipeline:7]')
+    assert '[pipeline:7] Using "a_vars["myvar"]" is deprecated. Use "VARS.myvar" instead.' in caplog.text
 
 
 @patch('automatix.config.VERSION', '2.1.5')