+++ automatix/environment.py
@@ -4,29 +4,43 @@ from logging import getLogger
 from .config import init_logger
 
 
+class AttributedDict(dict):
+    def __getattr__(self, key: str):
+        if key in self:
+            return self[key]
+        raise AttributeError
+
+    def __hasattr__(self, key: str):
+        return key in self
+
+    def __setattr__(self, key: str, value):
+        self[key] = value
+
+
 class PipelineEnvironment:
     def __init__(
             self,
-            name: str,
             config: dict,
-            systems: dict,
+            script: dict,
             variables: dict,
-            imports: list,
-            batch_mode: bool,
-            batch_items_count: int,
             batch_index: int,
             cmd_args: Namespace,
     ):
-        self.name = name
         self.config = config
-        self.systems = systems
-        self.vars = variables
-        self.imports = imports
-        self.batch_mode = batch_mode
-        self.batch_items_count = batch_items_count
+        self.script = script
+        self.vars = AttributedDict(variables)
         self.batch_index = batch_index
         self.cmd_args = cmd_args
+
+        self.name = script['name']
+        self.script_file_path = script['_script_file_path']
+        self.systems = script.get('systems', {})
+        self.imports = script.get('imports', [])
+        self.batch_mode = script.get('_batch_mode', False)
+        self.batch_items_count = script.get('_batch_items_count', 1)
+
         self.LOG = None
+        self.auto_file = None
 
         # This will be set at runtime
         self.command_count = None